All tests pass That means if I hook it up to the UI it should work perfectly first time. Totally confident.

+++ app/src/main/java/uk/co/whitetigergames/devtest_android/PublisherList.java
@@ -26,7 +26,7 @@ public class PublisherList implements IPublisherList
     }
 
     @Override
-    public long GetCount(String name)
+    public int GetCount(String name)
     {
         if(publishers.containsKey(name))
         {
