+++ app/src/main/java/uk/co/whitetigergames/devtest_android/interfaces/IPublisherList.java
@@ -7,5 +7,5 @@ public interface IPublisherList
 {
     void RecordInstance(String name);
 
-    long GetCount(String name);
+    int GetCount(String name);
 }
